Moved model/mesh/import db types into their own files
